% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{BoldDHSC-package}
\alias{BoldDHSC}
\alias{BoldDHSC-package}
\title{BOLD basic functions}
\description{
A package which contains basic functions to be often used by BOLD analysts
}
\keyword{internal}
