% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_completeness_heatmap.R
\name{plot_completeness_heatmap}
\alias{plot_completeness_heatmap}
\title{Create a factor defined heatmap of completeness}
\usage{
plot_completeness_heatmap(df, factor_col)
}
\arguments{
\item{df}{A dataframe}

\item{factor_col}{The name of the column which has the factor values in to be compared, given in quotations}
}
\value{
A heatmap as a ggplot2 object
}
\description{
The function creates a summary of missingness in a dataframe, split by a pareticular factor column
}
\examples{
plot_completeness_heatmap(airquality,"Month")

}
