% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_excel_to_csvs.R
\name{split_excel_to_csvs}
\alias{split_excel_to_csvs}
\title{Create multiple csvs from a fixed format excel document}
\usage{
split_excel_to_csvs(excel_path)
}
\arguments{
\item{excel_path}{The filepath of the input xlsx file}
}
\value{
Multiple csv documents which are saved in the working directory. The names of the csvs is a concatenation of the main xlsx file name and the sheet names.
}
\description{
The function creates multiple csv files from a single excel document
It is intended to only be used on the pre-defined data dictionary format and will break if the input file isnt in this form
}
\examples{
split_excel_to_csvs("Example_workbook.xlsx")
}
