% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\title{Create an Excel data dictionary from a dataframe}
\usage{
metadata(df, output_file)
}
\arguments{
\item{df}{A dataframe, ideally of less than 50 columns}

\item{output_file}{The filepath of the output xlsx file}
}
\value{
An xlsx document which is saved in the working directory. The first sheet is a summary of all of the columns in the data, and the second sheet is a lookup for each of the factors.
}
\description{
The function creates an excel summary of the dataframe, including missingness, example values, type of column.
It also requires the user to input the description for each column.
This can either be done in-line as the function runs, or can be uploaded as part of a reference date sheet
The file outputted also contains a sheet for the user to explicitly add value to factor columns.
}
\examples{
metadata(mtcars,"Example_output.xlsx")

}
